work only on has association
